Introduce flag for length access after push/unshift
For these methods, `Array` itself immediately gets the `.length` to
return after invoking them. Thus, if we are reading `.length`, it may
be a normal user-triggered read, or it may be a read triggered by Array
itself. In the latter case, it is because we have just done `.push()`
or `.unshift()`; in that case it is safe *not* to treat reading
`.length` as a *read* operation, since calling `.push()` or
`.unshift()` cannot otherwise be part of a "read" operation safely, and
if done during an *existing* read (e.g. if the user has already checked
`.length` *prior* to this), that will still trigger the assertion.

Accordingly, set the new flag to `true` when calling `.push()` or
`.unshift()`, and then clear it during the immediately-following
`.length` check.